refactor: `useSaveAudio` logic (part 2)
- Add missing check on `releaseYear` when creating a new album
- Move out the dedicated "indexing" logic from the hook
  - When updating a track, we let the track artwork deletion get handled by `cleanUpArtwork()`
